v1.0.0

+++ src/main/java/com/stlapp/stlappback/servicios/impl/UsuarioServiceImpl.java
@@ -1,15 +1,19 @@
 package com.stlapp.stlappback.servicios.impl;
 
 import com.stlapp.stlappback.excepciones.UsuarioFoundException;
+import com.stlapp.stlappback.excepciones.UsuarioNotFoundException;
 import com.stlapp.stlappback.modelos.Usuario;
 import com.stlapp.stlappback.modelos.UsuarioRol;
 import com.stlapp.stlappback.repositorios.RolRepository;
 import com.stlapp.stlappback.repositorios.UsuarioRepository;
 import com.stlapp.stlappback.servicios.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -20,6 +24,9 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Autowired
     private RolRepository rolRepository;
+
+    @Autowired
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Override
     public Usuario guardarUsuario(Usuario usuario, Set<UsuarioRol> usuarioRoles) throws Exception {
@@ -38,17 +45,73 @@ public class UsuarioServiceImpl implements UsuarioService {
     }
 
     @Override
-    public Usuario obtenerUsuario(String username) {
-        return usuarioRepository.findByUsername(username);
+    public ResponseEntity<Usuario> obtenerUsuario(Long usuarioId) throws Exception {
+
+        Usuario usuarioLocal = usuarioRepository.findById(usuarioId)
+                .orElseThrow(() -> new UsuarioNotFoundException("No se encuentra el usuario"));
+
+        return ResponseEntity.ok(usuarioLocal);
     }
 
     @Override
-    public void eliminarUsuario(Long usuarioId) {
-        usuarioRepository.deleteById(usuarioId);
+    public void eliminarUsuario(Long usuarioId) throws Exception {
+
+        Optional<Usuario> usuarioLocal = usuarioRepository.findById(usuarioId);
+        if(!usuarioLocal.isPresent()){
+            System.out.println("El usuario no está registrado");
+            throw new UsuarioNotFoundException("El usuario no existe en la bbdd");
+        }else{
+            usuarioRepository.deleteById(usuarioId);
+        }
     }
 
     @Override
     public List<Usuario> obtenerUsuarios(){
         return usuarioRepository.findAll();
     }
+
+    @Override
+    public ResponseEntity<Usuario> actualizarUsuario(Long usuarioId, Usuario usuario, Set<UsuarioRol> usuarioRoles) throws Exception{
+
+        Usuario usuarioLocal = usuarioRepository.findById(usuarioId)
+                .orElseThrow(() -> new UsuarioNotFoundException("El usuario no existe en la bbdd"));
+
+        if(usuario.getPerfil()!=null){
+            usuarioLocal.setPerfil(usuario.getPerfil());
+        }
+        if(usuario.getNombre()!=null){
+            usuarioLocal.setNombre(usuario.getNombre());
+        }
+        if(usuario.getApellidos()!=null){
+            usuarioLocal.setApellidos(usuario.getApellidos());
+        }
+        if(usuario.getEmail()!=null){
+            usuarioLocal.setEmail(usuario.getEmail());
+        }
+        if(usuario.getTelefono()!=null){
+            usuarioLocal.setTelefono(usuario.getTelefono());
+        }
+        if(usuario.getUsername()!=null) {
+            usuarioLocal.setUsername(usuario.getUsername());
+        }
+        if(usuario.getPassword()!=null){
+            usuarioLocal.setPassword(this.bCryptPasswordEncoder.encode(usuario.getPassword()));
+        }
+
+        if(usuarioRoles!=null){
+            for(UsuarioRol usuarioRol: usuarioRoles){
+                rolRepository.save(usuarioRol.getRol());
+            }
+            usuario.getUsuarioRoles().addAll(usuarioRoles);
+        }
+        if(usuario.getUsuarioRoles()!=null){
+            usuarioLocal.setUsuarioRoles(usuario.getUsuarioRoles());
+        }
+
+        Usuario usuarioActualizado = usuarioRepository.save(usuarioLocal);
+
+        return ResponseEntity.ok(usuarioActualizado);
+
+    }
+
 }
